+++ src/main/java/com/marginallyclever/makelangelo/donatelloimpl/nodes/turtle/PathImageMask.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.makelangelo.donatelloimpl.nodes;
+package com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle;
 
 import com.marginallyclever.convenience.Clipper2D;
 import com.marginallyclever.convenience.LineCollection;
