+++ src/main/java/com/marginallyclever/makelangelo/donatelloimpl/nodes/turtle/PatternOnPath.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.makelangelo.donatelloimpl.nodes;
+package com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle;
 
 import com.marginallyclever.convenience.Point2D;
 import com.marginallyclever.donatello.ports.InputInt;
