+++ src/main/java/com/marginallyclever/makelangelo/donatelloimpl/nodes/turtle/TurtleToBufferedImage.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.makelangelo.donatelloimpl.nodes;
+package com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle;
 
 import com.marginallyclever.makelangelo.donatelloimpl.ports.InputTurtle;
 import com.marginallyclever.donatello.ports.OutputImage;
