+++ src/main/java/com/marginallyclever/makelangelo/donatelloimpl/nodes/turtle/shapes/NGon.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.makelangelo.donatelloimpl.nodes.shapes;
+package com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle.shapes;
 
 import com.marginallyclever.donatello.ports.InputDouble;
 import com.marginallyclever.donatello.ports.InputInt;
