+++ src/main/java/module-info.java
@@ -35,7 +35,7 @@ module com.marginallyclever.makelangelo {
     exports com.marginallyclever.makelangelo.paper;
     exports com.marginallyclever.makelangelo.donatelloimpl to com.marginallyclever.nodegraphcore;
     exports com.marginallyclever.makelangelo.donatelloimpl.nodes to com.marginallyclever.nodegraphcore;
-    exports com.marginallyclever.makelangelo.donatelloimpl.nodes.shapes to com.marginallyclever.nodegraphcore;
+    exports com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle.shapes to com.marginallyclever.nodegraphcore;
     exports com.marginallyclever.convenience.log to ch.qos.logback.core;
 
     opens com.marginallyclever.convenience;
@@ -48,6 +48,7 @@ module com.marginallyclever.makelangelo {
     opens com.marginallyclever.convenience.log to ch.qos.logback.core;
     opens com.marginallyclever.makelangelo.preview;
     exports com.marginallyclever.makelangelo.donatelloimpl.ports to com.marginallyclever.nodegraphcore;
+    exports com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle to com.marginallyclever.nodegraphcore;
 
     // A Java module that wants to implement a service interface from a service interface module must:
     // - Require the service interface module in its own module descriptor.
